new workaround for ios and android
The previous workaround did not work on
Android because the event explicitOriginalTarget
does not exist on Android.

The new workaround suppreses the toggle
change response when the change is programmatic by introducing a check
in the change response handler.

A setTimeout of time 0 ms is used to ensure
that the check is reset only after the change
event is fired and handled by Angular.

Capacitor is added to quickly build and check
the solutionon iOS and Android.